Implement proper model usage and event logging
- Replace DTOs with actual model structs in handlers
- Fix type mismatches between database schema and models
- Implement EntryEvent model for tracking entry changes
- Add event logging to create_entry and update_entry functions
- Create events handler with GET endpoints for entry history
- Add proper type overrides for sqlx queries
- Remove dead code suppressions from model structs
- Fix all TypeScript and Rust compilation issues

+++ src/models/entry.rs
@@ -44,7 +44,7 @@ pub struct Attestation {
     pub quote_latin: Option<String>,
     pub quote_trans_id: Option<String>,
     pub quote_trans_en: Option<String>,
-    pub confidence: Option<f32>,
+    pub confidence: Option<f64>, // Changed to f64 to match NUMERIC(3,2)
     pub source_rank: i32,
     pub created_at: DateTime<Utc>,
 }
